Barcode/Barcode/requirements.txt
opencv-python
numpy
Pillow
pyzbar
pylibdmtx